==> 03_plot_ONI.py <==
#-----------------------------------------------------------------------------------------#
import sys
sys.path.append('./Libs') 
import utilities as U
#-----------------------------------------------------------------------------------------#
import pandas as pd
import os
#-----------------------------------------------------------------------------------------#

# NOTE Data loading
data = pd.read_csv('data_03/ONI.txt', delimiter='\s+')
output_folder = 'figures'

#-----------------------------------------------------------------------------------------#

os.makedirs(output_folder, exist_ok=True)

#-----------------------------------------------------------------------------------------#

# NOTE Plot data
U.plot_ONI(data, output_folder)

#-----------------------------------------------------------------------------------------#
